Processing modularisation
